refactor: fix type errors, improve file structure and enhance functionalities
