+++ src/components/sidebar/nav-main.tsx
@@ -2,8 +2,8 @@
 
 import { Link } from "@tanstack/react-router";
 import { ChevronRight } from "lucide-react";
-import { getRequestTypeClass } from "@/utils/request-type";
-import type { SidebarCollection } from "@/types/sidebar";
+import { getRequestTypeClass } from "@/utils/api/request-type";
+import type { SidebarCollection } from "@/types/components/sidebar";
 
 import {
 	Badge,
