+++ src/components/ui/field-renderer.tsx
@@ -1,5 +1,5 @@
 import type React from "react";
-import { getFieldComponent } from "@/utils/field-registry";
+import { getFieldComponent } from "@/utils/fields/field-registry";
 import type { ParameterObject } from "@/types/api/openapi";
 import UnsupportedField from "./fields/unsupported-field";
