+++ src/types/api/loader.ts
@@ -1,0 +1,8 @@
+/**
+ * Arguments for the loader function.
+ */
+export interface LoaderArgs {
+	params: {
+		api_url: string;
+	};
+}
