+++ src/types/api/openapi/index.ts
@@ -1,0 +1,6 @@
+export * from "./core";
+export * from "./paths";
+export * from "./operations";
+export * from "./parameters";
+export * from "./responses";
+export * from "./schemas";
