+++ src/types/api/openapi/parameters.ts
@@ -1,0 +1,11 @@
+import type { ReferenceObject, SchemaObject } from "./schemas";
+
+export type ParameterOrRef = ParameterObject | ReferenceObject;
+
+export interface ParameterObject {
+	name: string;
+	in: "path" | "query" | "header" | "cookie" | "body" | "response";
+	required?: boolean;
+	schema?: SchemaObject;
+	description?: string;
+}
